# backend/proveedores-service/src/swagger.yaml
openapi: 3.0.0
info:
  title: Proveedores Service API
  description: API para gestión de proveedores
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Servidor de desarrollo
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Servicio saludable
  /api/proveedores:
    get:
      summary: Obtener todos los proveedores
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Lista de proveedores
    post:
      summary: Crear nuevo proveedor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                ruc:
                  type: string
                email:
                  type: string
      responses:
        "201":
          description: Proveedor creado
  /api/proveedores/{id}:
    get:
      summary: Obtener proveedor por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Proveedor encontrado
    put:
      summary: Actualizar proveedor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Proveedor actualizado
    delete:
      summary: Eliminar proveedor
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Proveedor eliminado
  /api/proveedores/buscar/{termino}:
    get:
      summary: Buscar proveedores
      parameters:
        - in: path
          name: termino
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resultados de búsqueda
  /api/proveedores/estado/{estado}:
    get:
      summary: Obtener proveedores por estado
      parameters:
        - in: path
          name: estado
          required: true
          schema:
            type: string
            enum: [ACTIVO, INACTIVO, SUSPENDIDO]
      responses:
        "200":
          description: Proveedores encontrados